+++ ruthless-plugin/src/main/java/dev/lajoscseppento/ruthless/plugin/RuthlessJavaGradlePluginPlugin.java
@@ -4,14 +4,24 @@ import dev.lajoscseppento.ruthless.plugin.impl.AbstractProjectPlugin;
 import dev.lajoscseppento.ruthless.plugin.impl.RuthlessJavaBasePlugin;
 import java.util.Arrays;
 import java.util.List;
+import lombok.NonNull;
+import org.gradle.api.NamedDomainObjectProvider;
 import org.gradle.api.Plugin;
 import org.gradle.api.Project;
-import org.gradle.api.tasks.SourceSet;
-import org.gradle.api.tasks.testing.Test;
+import org.gradle.api.plugins.JavaPlugin;
+import org.gradle.api.plugins.jvm.JvmTestSuite;
+import org.gradle.api.plugins.jvm.JvmTestSuiteTarget;
 import org.gradle.language.base.plugins.LifecycleBasePlugin;
+import org.gradle.plugin.devel.GradlePluginDevelopmentExtension;
 import org.gradle.plugin.devel.plugins.JavaGradlePluginPlugin;
+import org.gradle.testing.base.TestingExtension;
 
 public class RuthlessJavaGradlePluginPlugin extends AbstractProjectPlugin {
+
+  private static final String TESTING_EXTENSION_NAME = "testing";
+
+  private GradlePluginDevelopmentExtension gradlePlugin;
+
   @Override
   protected List<Class<? extends Plugin<Project>>> requiredPlugins() {
     return Arrays.asList(RuthlessJavaBasePlugin.class, JavaGradlePluginPlugin.class);
@@ -19,18 +29,28 @@ public class RuthlessJavaGradlePluginPlugin extends AbstractProjectPlugin {
 
   @Override
   protected void apply() {
-    SourceSet functionalTestSourceSet = sourceSets.create("functionalTest");
-    gradlePlugin.testSourceSets(functionalTestSourceSet);
+    gradlePlugin = (GradlePluginDevelopmentExtension) extensions.getByName("gradlePlugin");
+    TestingExtension testing = (TestingExtension) extensions.getByName(TESTING_EXTENSION_NAME);
 
-    Test functionalTest =
-        tasks.create(
-            "functionalTest",
-            Test.class,
-            task -> {
-              task.setTestClassesDirs(functionalTestSourceSet.getOutput().getClassesDirs());
-              task.setClasspath(functionalTestSourceSet.getRuntimeClasspath());
-            });
+    NamedDomainObjectProvider<JvmTestSuite> functionalTestSuite =
+        testing
+            .getSuites()
+            .register("functionalTest", JvmTestSuite.class, this::configureFunctionalTestSuite);
 
-    tasks.getByName(LifecycleBasePlugin.CHECK_TASK_NAME).dependsOn(functionalTest);
+    tasks
+        .named(LifecycleBasePlugin.CHECK_TASK_NAME)
+        .configure(checkTask -> checkTask.dependsOn(functionalTestSuite));
+  }
+
+  private void configureFunctionalTestSuite(@NonNull JvmTestSuite functionalTest) {
+    gradlePlugin.testSourceSets(functionalTest.getSources());
+
+    functionalTest.getTargets().all(this::configureFunctionalTestSuiteTarget);
+  }
+
+  private void configureFunctionalTestSuiteTarget(@NonNull JvmTestSuiteTarget target) {
+    target
+        .getTestTask()
+        .configure(testTask -> testTask.shouldRunAfter(tasks.named(JavaPlugin.TEST_TASK_NAME)));
   }
 }