+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Declare Gradle Plugin functional test as JVM test suite
+
 ## [0.5.0] - 2022-07-21
 
 ### Added